match: caller_callee ARG-COUNT reconciliations (Framework B b3) - FAKE-VIEW 364->354
Same-class arg-count edges drained 8->0 (disasm-verified via ret N): CTriggerMgr::
ResetGroup 8->7 / CellDispatch 5->4, CGruntSpawnConfig::SpawnVoiceDriver 7->6 (the
reloc-masked trailing pads); value-size/int-ptr fixes (CObj23d90::Blit i16->i32,
StreamVoice::SetSource, CDDrawSubMgrLeaf::RemoveKeysEqual, CDDSurface::Lock,
CButeStore::ClearRecursive). MAX% 69.18 held. Remaining fake-view bulk is the
view!=owner hierarchy lane (Agent A) + 8 lower-ROI multi-cast edges.

+++ src/Gruntz/GameKeyHandler.cpp
@@ -70,7 +70,7 @@ struct CGroupSel {
 #include <Gruntz/FontConfig.h> // canonical CFontConfig (EndInput; non-virtual, cast-neutral)
 #include <Gruntz/GruntzMgr.h>  // canonical CGruntzMgr (score/run/finish helpers)
 struct CObj23d90 {
-    void Blit(i32 a, i32 b, i16 c, i16 d, i32 e);
+    void Blit(i32 a, i32 b, i32 c, i32 d, i32 e); // real @0x23d90 takes i32 args (caller widens the i16 word reads)
 };
 struct EngineLabelBacklog {
     i32 LoadExplosionSprites(i32 a, i32 b, i32 c, i32 d);
